Agregando estilos al formulario de post
